requirements.txt
ultralytics
opencv-python-headless 
numpy
tqdm
PyYAML
scikit-learn
redis
